requirements.txt
Flask
python-docx
PyPDF2
nltk
